requirements.txt
opencv-python
numpy
torch
torchvision
Pillow
python-chess
PyQt5
